+++ src/earthkit/time/__init__.py
@@ -7,7 +7,7 @@ from .sequence import (
     YearlySequence,
 )
 
-__version__ = "0.1.6"
+__version__ = "0.1.7"
 
 __all__ = [
     "__version__",